suicide

+++ app.py
@@ -59,31 +59,34 @@ def generate_map():
 @cache.memoize(timeout=6000)
 def crimeRateDistribution(df):
     time.sleep(10)
-    fig=px.histogram(df,x='Crime Rate',title="Crime Rate Distribution")
-    fig.write_html("templates/crimeRateDistribution.html")
-    print("Crime Rate Distribution Graph created")
-    shutil.copy('templates/crimeRateDistribution.html','static/crimeRateDistribution.html')
-    print("Crime Rate Distribution Graph Done !!")
+    if not os.path.exists('./templates/crime_map.html'):
+        fig=px.histogram(df,x='Crime Rate',title="Crime Rate Distribution")
+        fig.write_html("templates/crimeRateDistribution.html")
+        print("Crime Rate Distribution Graph created")
+        shutil.copy('templates/crimeRateDistribution.html','static/crimeRateDistribution.html')
+        print("Crime Rate Distribution Graph Done !!")
+        # return "templates/crimeRateDistribution.html"
     return "templates/crimeRateDistribution.html"
 
 @cache.memoize(timeout=6000)
 def TopCrimeHotSpot(df):
     time.sleep(15)
-    if 'City' not in df.columns or 'Crime Rate' not in df.columns:
-        return "Dataset don't have City and Crime rate columns"
-    cityCrimeRate=df.groupby('City_Name')['Crime Rate'].mean().reset_index()
-    cityCrimeRate=cityCrimeRate.sort_values(by='Crime Rate',ascending=False).head(10)
-    fig=px.bar(
-        cityCrimeRate,
-        x='City_Name',
-        y='Crime Rate',
-        title='Top 10 Crime Hotspots ',
-        color='Crime Rate',
-        color_continuous_scale='Reds'
-    )
-    print("Top 10 Done")
-    fig.write_html("templates/TopCrimeHotspot.html")
-    shutil.copy('templates/TopCrimeHotspot.html','static/TopCrimeHotspot.html')
+    if not os.path.exists('./templates/crime_map.html'):
+        if 'City' not in df.columns or 'Crime Rate' not in df.columns:
+            return "Dataset don't have City and Crime rate columns"
+        cityCrimeRate=df.groupby('City_Name')['Crime Rate'].mean().reset_index()
+        cityCrimeRate=cityCrimeRate.sort_values(by='Crime Rate',ascending=False).head(10)
+        fig=px.bar(
+            cityCrimeRate,
+            x='City_Name',
+            y='Crime Rate',
+            title='Top 10 Crime Hotspots ',
+            color='Crime Rate',
+            color_continuous_scale='Reds'
+        )
+        print("Top 10 Done")
+        fig.write_html("templates/TopCrimeHotspot.html")
+        shutil.copy('templates/TopCrimeHotspot.html','static/TopCrimeHotspot.html')
     return "templates/TopCrimeHotspot.html"
 
 @app.route('/',methods=['GET','POST'])
